revert cast, alphabetize / add deps

+++ performance_test/src/experiment_execution/analyze_runner.cpp
@@ -126,7 +126,7 @@ bool AnalyzeRunner::check_exit(std::chrono::steady_clock::time_point experiment_
   const double runtime_sec =
     std::chrono::duration<double>(std::chrono::steady_clock::now() - experiment_start).count();
 
-  if (static_cast<uint64_t>(runtime_sec) > m_ec.max_runtime()) {
+  if (runtime_sec > m_ec.max_runtime()) {
     std::cout << "Maximum runtime reached. Exiting." << std::endl;
     return true;
   } else {
